update diagonal lasers still not right but closer
