+++ index.html
@@ -15,7 +15,7 @@
         <br>• <span>arrow keys</span> control the ship •
         <br>• <span>space bar</span> fires bullets •
         <br>• <span>enter</span> to start game! •
-        <br>* ship invincible while red *
+        <br>• ship <span>invincible while red</span> •
       </p>
     </div>
 
@@ -25,7 +25,7 @@
     </div>
 
     <canvas id="game" width = "600" height = "600" style="display: none"></canvas>
-    
+
     <script src="lib/index.js"></script>
     <script src="main.bundle.js"></script>
     <audio src="lib/Carl-Sagan-Quotes.mp3" autoplay></audio>
