+++ lib/bullets.js
@@ -13,25 +13,25 @@ Bullet.prototype.draw = function (){
   var image = new Image();
   switch (this.direction) {
     case 'up':
-      image.src = "lib/images/lasers/laser-vertical.png";
+      image.src = "lib/images/lasers/lazer-vertical.png";
       break;
     case 'upRight':
       image.src = "lib/images/lasers/laser-45.png";
       break;
     case 'right':
-      image.src = "lib/images/lasers/laser-horizontal.png";
+      image.src = "lib/images/lasers/lazer-horizontal.png";
       break;
     case 'downRight':
       image.src = "lib/images/lasers/laser-315.png";
       break;
     case 'down':
-      image.src = "lib/images/lasers/laser-vertical.png";
+      image.src = "lib/images/lasers/lazer-vertical.png";
       break;
     case 'downLeft':
       image.src = "lib/images/lasers/laser-45.png";
       break;
     case 'left':
-      image.src = "lib/images/lasers/laser-horizontal.png";
+      image.src = "lib/images/lasers/lazer-horizontal.png";
       break;
     case 'upLeft':
       image.src = "lib/images/lasers/lazer-315.png";
