+++ lib/index.js
@@ -1,6 +1,6 @@
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
-var {Ship, drawLives} = require('./ship.js');
+var Ship = require('./ship.js');
 var Asteroid = require('./asteroids.js');
 var {Bullet, drawScore} = require('./bullets.js');
 var $  = require('jquery');
@@ -115,12 +115,36 @@ function runGame() {
   drawScore();
   generateAsteroids();
   bulletArray.forEach(function(bullet){
-    bullet.draw(context).move().collision(asteroidArray);
-  });
+      bullet.draw(context).move().collision(asteroidArray);
+    });
   shootBullets();
   ship.draw(gameCounter).move(rightKey,leftKey,upKey,downKey).collision(asteroidArray, gameCounter);
   asteroidArray.forEach(function (asteroid){
-    asteroid.draw().move();
-  });
+      asteroid.draw().move();
+    });
   gameCounter++;
 }
+
+function drawLives () {
+  context.fillStyle = 'white';
+  context.font = "16px Helvetica";
+  context.fillText("Lives: " + ship.lives, 8, 40);
+}
+
+$('#end-overlay').on('click', function () {restartGame();});
+
+function restartGame (){
+  location.reload();
+  // endGame();
+  // $('#game').hide();
+  // $('#end-overlay').hide();
+  // $('#start-overlay').show();
+  // $('button').focus();
+}
+//
+// function endGame(){
+//   context.clearRect(0,0, canvas.width, canvas.height);
+//   asteroidArray = [];
+//   bulletArray = [];
+//   gameCounter = 0;
+// }
