+++ lib/ship.js
@@ -9,6 +9,7 @@ function Ship (x,y, width, height, context) {
   this.height = height;
   this.direction = 'up';
   this.context = context;
+  this.lives = 3;
 }
 
 Ship.prototype.draw = function (counter) {
@@ -100,13 +101,14 @@ Ship.prototype.collision = function (asteroidArray, counter) {
       var asteroid = asteroidArray[i];
       if (this.x < asteroid.x + asteroid.width && this.x + this.width > asteroid.x &&
       this.y < asteroid.y + asteroid.height && this.height + this.y > asteroid.y) {
-        if (lives>1){
-          lives--;
+        if (this.lives>1){
+          this.lives--;
           this.x = 300;
           this.y = 300;
           deathCounter = counter;
         }
         else{
+          this.live--;
           deathCounter = 10000000000;
           showRestartScreen();
           return "game over";
@@ -122,22 +124,6 @@ function showRestartScreen(){
     $('#start-overlay').hide();
     $('#end-overlay').show(function(){$('#restart').focus();});
     $('#end-overlay').append(`<div class='info'>Your Score is: <span>${updateScore()}</span></div>`);
-    $('#end-overlay').on('click', function () {restartGame();});
-
 }
 
-function restartGame (){
-  // location.reload();
-  $('#game').remove();
-  $('body').prepend(`<canvas id="game" width = "600" height = "600" style="display: none"></canvas>`);
-  $('#start-overlay').show();
-  $('button').focus();
-}
-
-function drawLives () {
-  context.fillStyle = 'white';
-  context.font = "16px Helvetica";
-  context.fillText("Lives: " + lives, 8, 40);
-}
-
-module.exports = {Ship, drawLives};
+module.exports = Ship;
